Allunit tests that don#t require refactoring ready & refactored Database
In order to test the database, a double was used for the server called ServerDouble.  The double follows the same interface.

The database was tested, but onyl the methods that don't access the file system.  Methods that access the file system ( init() and persist() ) were decoupled from the others, and in the Manager they are called explicitly.

An interface for the database was created and a Database Double was created to test the Manager.  Currently there a few methods that test Manager.addCurrency but they are a mess.

Planning more unit tests (to cover exeptions) and polishing the Manager test.