Completed the tests for CurrencyManager
ChangedCurrency manager to use a factory that generates an iinstance of the CurrencyDatabase.  This is the best idea I have so far for Dependency Lookup, but I'll try to think of a better one.

ServerDouble and DatabaseDouble were moved back to the testing folder.  It makes more sense to keep the main program clean from clutter and test stuff separate.