Assignment 2:  Restarted it, imported dependencies & created test from feature
I found some issues with the previous project so I restarted it a Maven project from scratch, fixed the pom.xml file and added the scenarios.

All this because tests were not being created from scenarios but now I managed to do it for one scenario (its multiple lines).

It will all be attempted on amazon.co.uk