Assignment2: Cucumber works & feature file connected with stepDefs
All steps in the feature file have a corresponding empty test in AmazonUkStepdefs.

So the next step is doing stuff to the site.  I plan to add pageobjects & any helpful code as necessary.

Also added a file called SeparateTests which serves as my playground where I try new tests and see the results.  Not part of the assignment but I'll leave it there.