Assignment 2:  Wasted a whole day trying to make stuff work
At this point it should be as easy as getting web elements and doing things to them but it is far from that.
Amazon.co.uk does not just allow you to login, it sends a verification email with a code or asks for a mobile number for 2 factor aouthentication.  It's impossible to just login, so  I tried to work with AliExpress instead.

For some reason there is a uniquely identifiable webElement on the page that it just is Wuoting:  "Unable to find".  WTF?!