Assignment 2:  Part 1 completed successfully (see details)
All Scenarios tested and passed.  However, there is one that fails frequently & just found the cause.

The site does not let anyone order more than 9 of the same item, so the test to add 10 items to the cart fails because of that.

I will make the code tidier and mor maintainable and consider this section done.  Ghal erwieh!